requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
# TTS Dependencies
vieneu
soundfile
openai-whisper
numpy
lmdeploy

